+++ src/Filemanager/views/picker.blade.php
@@ -223,7 +223,7 @@
                                         {{ $file['name'] }}
                                     </a>
                                 </td>
-                                <td>{{ $file['mimeType'] or 'Unknown' }}</td>
+                                <td>{{ $file['mimeType'] ?? 'Unknown' }}</td>
                                 <td>{{ $file['modified']->format('j-M-y g:ia') }}</td>
                                 <td>{{ human_filesize($file['size']) }}</td>
                                 <td>